+++ docs/src/changelog.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 If your currently installed *Quality-time* version is v4.10.0 or older, please read the v5.0.0 deployment notes first.
 
+### Fixed
+
+- When measuring the duration of a running pipeline with GitLab as source, an error would be reported. Fixes [#8123](https://github.com/ICTU/quality-time/issues/8123).
+
 ### Added
 
 - If the documentation at Read-the-Docs has specific information on configuring a metric or source, point this out in Quality-time itself. Closes [#4445](https://github.com/ICTU/quality-time/issues/4445).